requirements.txt
asgiref==3.5.0
cffi==1.15.0
click==8.1.2
colorama==0.4.4
cryptography==37.0.1
Flask==2.1.1
Flask-Cors==3.0.10
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
opencv-python==4.5.5.64
pep8==1.7.1
Pillow==9.0.1
pycparser==2.21
pygifsicle==1.0.5
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
Werkzeug==2.1.1
zipp==3.8.0